requirements.txt
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
flask-view-counter==0.1.2
Flask-WTF==0.14.3
geojson==2.5.0
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
password-strength==0.0.3.post2
Pillow==8.0.1
pycparser==2.20
pyowm==3.1.1
pyshorteners==1.0.1
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
rfc3986==1.4.0
setuptools==49.2.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
